+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed some issues when multi editing HDRenderPipelineAsset
 - Fixed emission node breaking the main shader graph preview in certain conditions.
 - Fixed checkout of baked probe asset when baking probes.
+- Fixed some resources to handle more than 2 instanced views for XR
 
 ### Changed
 - DensityVolume scripting API will no longuer allow to change between advance and normal edition mode